Add Null VM tests and merge  some tests to decrease compile time. This depends on https://github.com/proxy-wasm/proxy-wasm-cpp-host/pull/31
