+++ source/extensions/common/wasm/BUILD
@@ -2,7 +2,6 @@ load(
     "//bazel:envoy_build_system.bzl",
     "envoy_cc_library",
     "envoy_package",
-    "envoy_proto_library",
 )
 
 licenses(["notice"])  # Apache 2
@@ -70,7 +69,6 @@ envoy_cc_library(
         "wasm_vm.cc",
     ],
     deps = [
-        ":declare_property_cc_proto",
         ":wasm_hdr",
         ":wasm_interoperation_lib",
         "//external:abseil_base",
@@ -82,6 +80,7 @@ envoy_cc_library(
         "//source/common/http:message_lib",
         "//source/common/http:utility_lib",
         "//source/common/tracing:http_tracer_lib",
+        "//source/extensions/common/wasm/ext:declare_property_cc_proto",
         "//source/extensions/common/wasm/ext:envoy_null_vm_wasm_api",
         "//source/extensions/filters/common/expr:context_lib",
         "@com_google_cel_cpp//eval/eval:field_access",
@@ -95,9 +94,4 @@ envoy_cc_library(
         "@envoy_api//envoy/extensions/wasm/v3:pkg_cc_proto",
         "@proxy_wasm_cpp_host//:lib",
     ],
-)
-
-envoy_proto_library(
-    name = "declare_property",
-    srcs = ["declare_property.proto"],
 )
