+++ source/extensions/common/wasm/ext/BUILD
@@ -27,15 +27,33 @@ envoy_cc_library(
         "envoy_null_plugin.h",
         "envoy_proxy_wasm_api.h",
     ],
+    deps = [
+        ":declare_property_cc_proto",
+        "//source/common/grpc:async_client_lib",
+    ],
 )
 
+# NB: this target is compiled to Wasm. Hence the generic rule.
 cc_library(
     name = "envoy_proxy_wasm_api_lib",
     srcs = ["envoy_proxy_wasm_api.cc"],
     hdrs = ["envoy_proxy_wasm_api.h"],
     tags = ["manual"],
     deps = [
+        ":declare_property_cc_proto",
         "@proxy_wasm_cpp_sdk//:proxy_wasm_intrinsics",
     ],
     alwayslink = 1,
+)
+
+# NB: this target is compiled both to native code and to Wasm. Hence the generic rule.
+proto_library(
+    name = "declare_property_proto",
+    srcs = ["declare_property.proto"],
+)
+
+# NB: this target is compiled both to native code and to Wasm. Hence the generic rule.
+cc_proto_library(
+    name = "declare_property_cc_proto",
+    deps = [":declare_property_proto"],
 )
