+++ source/extensions/common/wasm/ext/declare_property.proto
