--- test/extensions/filters/http/wasm/test_data/null_plugin/BUILD
@@ -1,28 +1,0 @@
-licenses(["notice"])  # Apache 2
-
-load(
-    "//bazel:envoy_build_system.bzl",
-    "envoy_cc_library",
-    "envoy_package",
-)
-
-envoy_package()
-
-envoy_cc_library(
-    name = "plugin",
-    srcs = [
-        "plugin.cc",
-        "null_plugin.cc",
-    ],
-    copts = ["-DNULL_PLUGIN=1"],
-    deps = [
-        "//external:abseil_node_hash_map",
-        "//source/common/common:assert_lib",
-        "//source/common/common:c_smart_ptr_lib",
-        "//source/common/protobuf",
-        "//source/extensions/common/wasm:declare_property_cc_proto",
-        "//source/extensions/common/wasm:wasm_hdr",
-        "//source/extensions/common/wasm:wasm_lib",
-        "//source/extensions/common/wasm:well_known_names",
-    ],
-)
